requirements.txt
requests>=2.31.0
pandas>=2.0.0
streamlit>=1.32.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
scikit-learn>=1.0.0
beautifulsoup4>=4.12.2
selenium>=4.10.0
webdriver-manager>=3.8.6
python-dotenv>=1.0.0
